% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_rprof}
\alias{parse_rprof}
\title{Parse Rprof output file for use with profvis}
\usage{
parse_rprof(path = "Rprof.out", expr_source = NULL)
}
\arguments{
\item{path}{Path to the \code{\link{Rprof}} output file.}

\item{expr_source}{If any source refs in the profiling output have an empty
filename, that means they refer to code executed at the R console. This
code can be captured and passed (as a string) as the \code{expr_source}
argument.}
}
\description{
Parse Rprof output file for use with profvis
}
