% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pause.R
\name{pause}
\alias{pause}
\title{Pause an R process}
\usage{
pause(seconds)
}
\arguments{
\item{seconds}{Number of seconds to pause. Note that \code{pause} only has
1-second precision.}
}
\description{
This function pauses an R process for some amount of time. It differs from
\code{\link{Sys.sleep}} in that time spent in \code{pause} will show up in
profiler data. Another difference is that \code{pause} uses up 100% of a CPU,
whereas \code{Sys.sleep} does not.
}

