% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prof.R
\name{prof}
\alias{prof}
\title{Run an R expression and record profiling information}
\usage{
prof(expr, interval = 0.01, prof_file = NULL)
}
\arguments{
\item{expr}{Code to profile.}

\item{interval}{Interval for profiling samples, in seconds. Values less than
0.005 (5 ms) will probably not result in accurate timings}

\item{prof_file}{Name of an Rprof output file in which to save profiling
  data. If \code{NULL} (the default), a temporary file will be used and
  automatically removed when the function exits.}
}
\description{
Run an R expression and record profiling information
}
\seealso{
\code{\link{Rprof}} for more information about how the profiling
  data is collected.
}

