% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profvis-package.R
\docType{package}
\name{profvis-package}
\alias{profvis-package}
\title{profvis: Interactive Visualizations for Profiling R Code}
\description{
Interactive visualizations for profiling R code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://profvis.r-lib.org}
  \item \url{https://github.com/r-lib/profvis}
  \item Report bugs at \url{https://github.com/r-lib/profvis/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@posit.co}

Authors:
\itemize{
  \item Winston Chang
  \item Javier Luraschi
  \item Timothy Mastny
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
  \item  jQuery Foundation (jQuery library) [copyright holder]
  \item  jQuery contributors (jQuery library; authors listed in inst/htmlwidgets/lib/jquery/AUTHORS.txt) [contributor, copyright holder]
  \item Mike Bostock (D3 library) [contributor, copyright holder]
  \item  D3 contributors (D3 library) [contributor]
  \item Ivan Sagalaev (highlight.js library) [contributor, copyright holder]
}

}
\keyword{internal}
