% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_module.R
\name{profvis_server}
\alias{profvis_server}
\title{Profvis Server for Shiny Apps}
\usage{
profvis_server(input, output, session, dir = ".")
}
\arguments{
\item{input, output, session}{Arguments used by
\code{\link[shiny]{callModule}}.}

\item{dir}{Output directory to save Rprof files.}
}
\description{
Provides the server for the Profvis module in Shiny apps. This module creates
a profvis interface to profile parts of a shiny app and save the profvis or
Rprof output. Used with \code{\link{profvis_ui}}. This function requires the
\code{shiny} package.
}
\examples{
if(interactive()) {
  shinyApp(
    fluidPage(
      plotOutput("plot"),
      actionButton("new", "New plot"),
      profvis_ui("profiler")
    ),
    function(input, output, session) {
      callModule(profvis_server, "profiler")

      output$plot <- renderPlot({
        input$new
        ggplot(diamonds, aes(carat, price)) + geom_point()
      })
    }
  )
}

}
