% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profvis.R
\name{print.profvis}
\alias{print.profvis}
\title{Print a profvis object}
\usage{
\method{print}{profvis}(x, ..., viewer = FALSE)
}
\arguments{
\item{viewer}{If \code{FALSE} (the default), display in an external web
browser. If \code{TRUE}, attempt to display in the RStudio viewer pane.
This can be useful for publishing profvis visualizations.}
}
\description{
Print a profvis object
}

