% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profvis.R
\name{print.profvis}
\alias{print.profvis}
\title{Print a profvis object}
\usage{
\method{print}{profvis}(x, ..., width = NULL, height = NULL, split = NULL, aggregate = NULL)
}
\arguments{
\item{x}{The object to print.}

\item{...}{Further arguments to passed on to other print methods.}

\item{width}{Width of the htmlwidget.}

\item{height}{Height of the htmlwidget}

\item{split}{Orientation of the split bar: either \code{"h"} (the default) for
horizontal or \code{"v"} for vertical.}

\item{aggregate}{If \code{TRUE}, the profiled stacks are aggregated by
name. This makes it easier to see the big picture. Set your own
global default for this argument with \code{options(profvis.aggregate = )}.}
}
\description{
Print a profvis object
}
