% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_module.R
\name{profvis_ui}
\alias{profvis_ui}
\title{Profvis UI for Shiny Apps}
\usage{
profvis_ui(id)
}
\arguments{
\item{id}{Output id from \code{profvis_server}.}
}
\description{
Provides the UI for the Profvis module in Shiny apps. This module creates a
profvis interface to profile parts of a shiny app and save the profvis or
Rprof output. Used with \code{\link{profvis_server}}. This function requires
the \code{shiny} package.
}
\examples{
if(interactive()) {
  shinyApp(
    fluidPage(
      plotOutput("plot"),
      actionButton("new", "New plot"),
      profvis_ui("profiler")
    ),
    function(input, output, session) {
      callModule(profvis_server, "profiler")

      output$plot <- renderPlot({
        input$new
        ggplot(diamonds, aes(carat, price)) + geom_point()
      })
    }
  )
}

}
