% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profvis.R
\name{profvisOutput}
\alias{profvisOutput}
\title{Widget output function for use in Shiny}
\usage{
profvisOutput(outputId, width = "100\%", height = "600px")
}
\arguments{
\item{outputId}{Output variable for profile visualization.}

\item{width}{Width of the htmlwidget.}

\item{height}{Height of the htmlwidget}
}
\description{
Widget output function for use in Shiny
}

