% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/profvis.R
\name{renderProfvis}
\alias{renderProfvis}
\title{Widget render function for use in Shiny}
\usage{
renderProfvis(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

